backend/requirements.txt
anyio==4.11.0
blinker==1.9.0
CacheControl==0.14.3
cachetools==6.2.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cryptography==46.0.2
firebase_admin==7.1.0
Flask==3.1.2
flask-cors==6.0.1
google-api-core==2.25.2
google-auth==2.41.1
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.4.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-status==1.75.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
msgpack==1.1.1
proto-plus==1.26.1
protobuf==6.32.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
python-dotenv==1.1.1
requests==2.32.5
rsa==4.9.1
sniffio==1.3.1
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
